teamB_section8/requirements.txt
# FastAPI本体
fastapi==0.110.2
uvicorn[standard]==0.29.0

# DB/ORM
SQLAlchemy==2.0.30
alembic==1.13.1

# SQLite用ドライバ
aiosqlite==0.20.0

# LINE BOT SDK
line-bot-sdk==3.10.1

# OpenAI API
openai==1.30.1

# 環境変数管理
python-dotenv==1.0.1

# ログ
loguru==0.7.2

# 型・バリデーション
pydantic==2.7.1

# Windows行末変換用（イメージビルド時のみ）
dos2unix 